server/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
databases==0.9.0
Deprecated==1.2.18
fastapi==0.115.12
googleapis-common-protos==1.70.0
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.6.1
Mako==1.3.10
MarkupSafe==3.0.2
objgraph==3.6.2
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==25.0
protobuf==5.29.5
psutil==7.0.0
psycopg2==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
zipp==3.22.0
